Web/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyjwt==2.9.0
pyotp==2.9.0
python-dotenv==1.0.1
pydantic-settings==2.3.4
SQLAlchemy==2.0.36
alembic==1.14.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.1
pytest==8.3.3
email-validator==2.2.0
PyMySQL==1.1.1
qrcode[pil]==7.4.2
